Add files via upload
made the anchor box become float, fixed minor issues before big changes

+++ utilis.py
@@ -151,7 +151,7 @@ def shortcut_layer_handling(module, index, layer, layer_type_dic):
 def yolo_layer_handling(module, index, layer, layer_type_dic):
     anchor_index = [int(x) for x in (layer["mask"].split(","))]
     anchors = re.split(',  |,', layer["anchors"])
-    anchors = np.reshape([int(x) for x in anchors], (9, 2))
+    anchors = np.reshape([float(x) for x in anchors], (9, 2))
     anchors = anchors[anchor_index]
     num_anchors = len(anchors)
     classes = int(layer["classes"])
